+++ filters/__init__.py
@@ -1,4 +1,5 @@
 from .VDM import VDM
 from .MrmrDiscreteFilter import MrmrDiscreteFilter
+from .SymmetricUncertainty import SymmetricUncertaintyFilter
 from .Filter import *
